requirements.txt
pymupdf==1.22.3
pdfplumber==0.9.0
pandas==2.0.3
requests==2.31.0
tqdm==4.66.1
faiss-cpu==1.7.4
numpy==1.24.3
google-generativeai==0.3.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
streamlit==1.24.0 